ai-development/requirements.txt
# Core
PyYAML==6.0.1

# Development
pytest>=7.0.0
